--- src/Projects/Modules/HorribleSubs/Clients/Module.HorribleSubs.Avalonia/PackagesInfo.json
@@ -1,4 +1,0 @@
-{
-  "assemblyFile": "Module.HorribleSubs.Avalonia.dll",
-  "entryType": "Module.HorribleSubs.Avalonia.Module"
-}
