--- src/Projects/Modules/HorribleSubs/Server/Module.HorribleSubs.Cida/Models/DownloadProgress.cs
@@ -1,8 +1,0 @@
-﻿namespace Module.HorribleSubs.Cida.Models
-{
-    public class DownloadProgress
-    {
-        public ulong DownloadedBytes { get; set; }
-        public ulong Size { get; set; }
-    }
-}
