--- src/Projects/Modules/HorribleSubs/Server/Module.HorribleSubs.Cida/Models/DownloadRequest.cs
@@ -1,11 +1,0 @@
-﻿namespace Module.HorribleSubs.Cida.Models
-{
-    public class DownloadRequest
-    {
-        public string BotName { get; set; }
-
-        public long PackageNumber { get; set; }
-
-        public string FileName { get; set; }
-    }
-}
