--- src/Projects/Modules/HorribleSubs/Server/Module.HorribleSubs.Cida/Models/SearchResult.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Module.HorribleSubs.Cida.Models
-{
-    public class SearchResult
-    {
-        public string BotName { get; set; }
-
-        public long PackageNumber { get; set; }
-
-        public long FileSize { get; set; }
-
-        public string FileName { get; set; }
-    }
-}
