--- src/Projects/Modules/HorribleSubs/Server/Module.HorribleSubs.Cida/Module.HorribleSubs.Cida.csproj
@@ -1,31 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <LangVersion>8</LangVersion>
-    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Grpc.Core.Api" Version="2.24.0" />
-    <PackageReference Include="Google.Protobuf" Version="3.9.1" />
-    <PackageReference Include="Grpc" Version="2.23.0" />
-    <PackageReference Include="Grpc.Tools" Version="2.23.0">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.0.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.0.1" />
-    <PackageReference Include="System.Text.Json" Version="4.7.0" />
-    <ProjectReference Include="..\..\..\..\Cida.Api\Cida.Api.csproj" />
-    <ProjectReference Include="..\..\..\..\IrcClient\IrcClient.csproj" />
-    <Protobuf Include="..\..\..\..\..\protos\Modules\horriblesubs.proto" Link="horriblesubs.proto" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="PackagesInfo.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
