--- src/Projects/Modules/HorribleSubs/Server/Module.HorribleSubs.Cida/PackagesInfo.json
@@ -1,7 +1,0 @@
-{
-  "id": "109F8F54-DE94-479A-840A-7B4EF0F284D2",
-  "name": "HorribleSubs",
-  "assemblyFile": "Module.HorribleSubs.Cida.dll",
-  "entryType": "Module.HorribleSubs.Cida.Module",
-  "clients": {}
-}
