+++ src/Projects/Modules/IrcAnime/Clients/Module.IrcAnime.Avalonia/Assets/SideBar.xaml
