+++ src/Projects/Modules/IrcAnime/Clients/Module.IrcAnime.Avalonia/Icons/Icons.xaml
