+++ src/Projects/Modules/IrcAnime/Clients/Module.IrcAnime.Avalonia/PackagesInfo.json
@@ -1,0 +1,4 @@
+{
+  "assemblyFile": "Module.IrcAnime.Avalonia.dll",
+  "entryType": "Module.IrcAnime.Avalonia.Module"
+}
