+++ src/Projects/Modules/IrcAnime/Clients/Module.IrcAnime.Avalonia/Views/IrcAnimeView.xaml.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Linq;
+using System.Reactive.Disposables;
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Controls.Primitives;
+using Avalonia.Markup.Xaml;
+using Avalonia.ReactiveUI;
+using Module.IrcAnime.Avalonia.ViewModels;
+using ReactiveUI;
+
+namespace Module.IrcAnime.Avalonia.Views
+{
+    public class IrcAnimeView : ReactiveUserControl<IrcAnimeViewModel>
+    {
+        public IrcAnimeView()
+        {
+            InitializeComponent();
+        }
+
+        private void InitializeComponent()
+        {
+            AvaloniaXamlLoader.Load(this);
+        }
+    }
+}
