+++ src/Projects/Modules/IrcAnime/Server/Module.IrcAnime.Cida/Models/DownloadProgress.cs
@@ -1,0 +1,8 @@
+﻿namespace Module.IrcAnime.Cida.Models
+{
+    public class DownloadProgress
+    {
+        public ulong DownloadedBytes { get; set; }
+        public ulong Size { get; set; }
+    }
+}
