+++ src/Projects/Modules/IrcAnime/Server/Module.IrcAnime.Cida/Models/DownloadRequest.cs
@@ -1,0 +1,11 @@
+﻿namespace Module.IrcAnime.Cida.Models
+{
+    public class DownloadRequest
+    {
+        public string BotName { get; set; }
+
+        public long PackageNumber { get; set; }
+
+        public string FileName { get; set; }
+    }
+}
