+++ src/Projects/Modules/IrcAnime/Server/Module.IrcAnime.Cida/Models/SearchResult.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Module.IrcAnime.Cida.Models
+{
+    public class SearchResult
+    {
+        public string BotName { get; set; }
+
+        public long PackageNumber { get; set; }
+
+        public long FileSize { get; set; }
+
+        public string FileName { get; set; }
+    }
+}
